refactor: create stats comp, convert about to tsx
