+++ src/pages/Market/About/About.module.scss
@@ -26,50 +26,6 @@
   }
 }
 
-.card__header--big {
-  text-align: left;
-  position: relative;
-  color: #0b0f1c;
-  font-weight: 700;
-  font-size: 30px;
-  margin-bottom: 25px;
-
-  &:after {
-    position: absolute;
-    content: "";
-    height: 2px;
-    /* adjust this to move up and down. */
-    bottom: -5px;
-
-    /* center - use with adjusting width   */
-    margin: 0 auto;
-    left: 0;
-    width: 85%;
-    background: black;
-  }
-
-  @include lg {
-    font-size: 36px;
-  }
-}
-
-.card__header--small {
-  text-align: left;
-  color: #0b0f1c;
-  font-weight: 500;
-  font-size: 24px;
-}
-
-.row__container {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 15px;
-
-  @include lg {
-    margin-bottom: 25px;
-  }
-}
-
 .row__number {
   color: #e0e0e0;
   font-size: 24px;
@@ -82,69 +38,4 @@
   font-weight: 400;
   text-align: left;
   width: 70%;
-}
-
-.column__container {
-  display: flex;
-  flex-direction: column;
-  text-align: left;
-  margin-bottom: 15px;
-  padding-left: 10px;
-}
-
-.icon__container {
-  height: 36px;
-  width: 36px;
-  margin-right: 10px;
-
-  img {
-    height: 100%;
-    width: 100%;
-  }
-}
-
-.link {
-  margin: auto 0;
-  color: #0b0f1c;
-  text-decoration: none;
-  position: relative;
-  font-size: 20px;
-
-  &:after {
-    position: absolute;
-    content: "";
-    height: 1px;
-    /* adjust this to move up and down. you may have to adjust the line height of the paragraph if you move it down a lot. */
-    bottom: 1px;
-
-    /* center - (optional) use with adjusting width   */
-    margin: 0 auto;
-    left: 0;
-    right: 0;
-    width: 100%;
-    background: #0b0f1c;
-
-    /* optional animation */
-    -o-transition: 0.5s;
-    -ms-transition: 0.5s;
-    -moz-transition: 0.5s;
-    -webkit-transition: 0.5s;
-    transition: 0.5s;
-  }
-  &:hover:after {
-    width: 80%;
-    background: blue;
-  }
-}
-
-.stat__header {
-  text-decoration: underline;
-  font-size: 24px;
-  margin-bottom: 5px;
-}
-
-.exchange-button {
-  margin-bottom: 20px !important;
-  text-transform: capitalize !important;
-  width: 85%;
 }
