+++ src/pages/Market/Stats/Stats.module.scss
@@ -1,0 +1,13 @@
+.column__container {
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  margin-bottom: 15px;
+  padding-left: 10px;
+}
+
+.stat__header {
+  text-decoration: underline;
+  font-size: 24px;
+  margin-bottom: 5px;
+}
